fix[dashboard]: reorganiza arquivos de função do dashboard
Refs: #47
